# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: NodelistDB API
  description: |
    NodelistDB is a comprehensive API for accessing historical FTN nodelist
    data: every FidoNet nodelist since 1986, other FTN networks (fsxNet, ...),
    pointlists (FTS-5002), and what the testing daemon found when it called
    each node.

    ## Features
    - **Node search**: find nodes by zone, net, node number, system name,
      location, sysop, capability flags or date range, in any network
    - **Historical data**: the complete history, change log and timeline of
      any node
    - **Multi-network**: `GET /api/networks` lists the imported FTN networks;
      most endpoints take `?domain=`
    - **Points**: pointlist snapshots under a boss, point history, imported
      pointlist issues
    - **Statistics**: per-date network statistics and the list of nodelist dates
    - **Reachability**: the testing daemon's results per node (BinkP, IFCICO,
      Telnet, FTP, VModem, over IPv4 and IPv6), trends and searches
    - **Software, hosting, PING/TRACE analytics**: distributions computed from
      those tests, and the netmail PING measurements
    - **Flag documentation**: a reference for FidoNet node flags

    ## Multi-network conventions
    A node address is only unique within one FTN network. Endpoints about one
    address resolve the network in this order: an explicit `?domain=`; the
    only network the address exists in; `fidonet` when it exists in several.
...
    as their descriptions say.

...

...

...
servers:
...
paths:
...
  /api/nodes/pstn:
    get:
      summary: Get PSTN Nodes
      description: |
        Every node in the latest nodelist with a usable phone number, for
        modem testing. Nodes marked dead via /api/modem/pstn-dead are
        excluded.
      operationId: getPSTNNodes
      tags:
        - Nodes
      parameters:
        - $ref: '#/components/parameters/DomainDefaultFidonet'
        - name: zone
          in: query
          description: Restrict to one zone
          schema:
            type: integer
        - name: limit
          in: query
          description: Maximum nodes to return (default 5000; values above 10000 are capped)
          schema:
            type: integer
            default: 5000
            maximum: 10000
      responses:
        '200':
          description: PSTN-capable nodes
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    type: array
                    items:
                      $ref: '#/components/schemas/PSTNNode'
                  count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/nodes/pstn/recent-success:
    get:
      summary: Get Recently Reached Phone Numbers
      description: |
        Phone numbers a modem test reached successfully within the window, so
        a caller can skip them.
      operationId: getRecentModemSuccessPhones
      tags:
        - Nodes
      parameters:
        - name: days
          in: query
          description: Look-back window in days (1-90)
          schema:
            type: integer
            default: 7
            minimum: 1
            maximum: 90
      responses:
        '200':
          description: Phone numbers reached in the window
          content:
            application/json:
              schema:
                type: object
                properties:
                  phones:
                    type: array
                    items:
                      type: string
                  count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/nodes/{zone}/{net}/{node}/history:
    get:
      summary: Get Node History
      description: Every nodelist entry of one address, oldest first, with the span it covers.
      operationId: getNodeHistory
      tags:
        - Nodes
      parameters:
        - $ref: '#/components/parameters/Zone'
        - $ref: '#/components/parameters/Net'
        - $ref: '#/components/parameters/Node'
        - $ref: '#/components/parameters/DomainResolved'
      responses:
        '200':
          description: Node history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AddressEnvelope'
                  - type: object
                    properties:
                      history:
                        type: array
                        items:
                          $ref: '#/components/schemas/Node'
                      count:
                        type: integer
                      first_date:
                        type: string
                        format: date-time
                      last_date:
                        type: string
                        format: date-time
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/nodes/{zone}/{net}/{node}/changes:
    get:
      summary: Get Node Changes
      description: |
        The change log of one address: one entry per nodelist date on which
        the node was added, removed or modified, with the fields that changed.
        An address with no history answers an empty list.
      operationId: getNodeChanges
      tags:
        - Nodes
      parameters:
        - $ref: '#/components/parameters/Zone'
        - $ref: '#/components/parameters/Net'
        - $ref: '#/components/parameters/Node'
        - $ref: '#/components/parameters/DomainResolved'
      responses:
        '200':
          description: Node changes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/AddressEnvelope'
                  - type: object
                    properties:
                      changes:
                        type: array
                        items:
                          $ref: '#/components/schemas/NodeChange'
                      count:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '503':
          $ref: '#/components/responses/BudgetExceeded'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    AddressEnvelope:
      type: object
      description: >
        The preamble of every response about one 3-D or 4-D address: which
        network answered and which networks the address exists in.
      properties:
        address:
          type: string
          example: "2:5001/100"
        domain:
          type: string
          example: fidonet
        available_domains:
          type: array
          items:
            type: string
          description: Always an array, possibly empty
...
    Error:
      type: object
      description: Every JSON error body
      properties:
        error:
          type: string
          description: What went wrong
        status:
          type: integer
          description: The HTTP status, repeated
        time:
          type: string
          format: date-time
...
    Node:
      type: object
      properties:
        zone:
          type: integer
        net:
          type: integer
        node:
          type: integer
        domain:
          type: string
          description: Omitted when empty.
        nodelist_date:
          type: string
          format: date-time
        day_number:
          type: integer
        system_name:
          type: string
        location:
          type: string
        sysop_name:
          type: string
        phone:
          type: string
        node_type:
          type: string
        region:
          type: integer
          nullable: true
          description: Omitted when empty.
        max_speed:
          type: integer
        is_cm:
          type: boolean
        is_mo:
          type: boolean
        has_inet:
          type: boolean
        flags:
          type: array
          items:
            type: string
        modem_flags:
          type: array
          items:
            type: string
        internet_config:
          type: object
          additionalProperties: true
          description: Omitted when empty.
        conflict_sequence:
          type: integer
        has_conflict:
          type: boolean
        fts_id:
          type: string
        raw_line:
          type: string
          description: Omitted when empty.
    NodeChange:
      type: object
      properties:
        date:
          type: string
          format: date-time
        day_number:
          type: integer
        change_type:
          type: string
        changes:
          type: object
          additionalProperties:
            type: string
        old_node:
          $ref: '#/components/schemas/Node'
          description: Omitted when empty.
        new_node:
          $ref: '#/components/schemas/Node'
          description: Omitted when empty.
...
    PSTNNode:
      type: object
      properties:
        zone:
          type: integer
        net:
          type: integer
        node:
          type: integer
        system_name:
          type: string
        location:
          type: string
        sysop_name:
          type: string
        phone:
          type: string
        phone_normalized:
          type: string
        is_cm:
          type: boolean
        nodelist_date:
          type: string
          format: date-time
        node_type:
          type: string
        max_speed:
          type: integer
        flags:
          type: array
          items:
            type: string
        modem_flags:
          type: array
          items:
            type: string
        is_pstn_dead:
          type: boolean
        pstn_dead_reason:
          type: string
          description: Omitted when empty.
...
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        A caller-specific API key. The server stores its SHA-256 hash in
        modem_api.callers; see the README for generating a pair.
  responses:
    BadRequest:
      description: Invalid or missing parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: No such record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    TooManyRequests:
      description: >
        Rate limited. The body is plain text and Retry-After says how many
        seconds to wait.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        text/plain:
          schema:
            type: string
    BudgetExceeded:
      description: The query outran its time budget; narrow it or retry later
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Zone:
      name: zone
      in: path
      required: true
      description: FTN zone number (FidoNet uses 1-6; other networks have their own)
      schema:
        type: integer
        minimum: 1
    Net:
      name: net
      in: path
      required: true
      description: Network number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
    Node:
      name: node
      in: path
      required: true
      description: Node number (0-65535)
      schema:
        type: integer
        minimum: 0
        maximum: 65535
...
    DomainResolved:
      name: domain
      in: query
      description: >
        FTN network. Omitted: the only network the address exists in, or
        fidonet when it exists in several.
      schema:
        type: string
        example: fsxnet
    DomainDefaultFidonet:
      name: domain
      in: query
      description: FTN network (defaults to fidonet)
      schema:
        type: string
        default: fidonet
...